Add initial store assets for PaperRescue app
- Create Play Store listing with app name, descriptions, category, and keywords.
- Add release notes for version 1.0.0 detailing features and capabilities.
- Include screenshot captions for Play Store with suggested order and descriptions.

+++ android/app/src/main/java/com/oldalexhub/paperrescue/scanner/DocumentOverlayView.kt
@@ -48,8 +48,8 @@ class DocumentOverlayView @JvmOverloads constructor(
         val offsetX = (width - sourceWidth * scale) / 2f
         val offsetY = (height - sourceHeight * scale) / 2f
 
-        fun mapX(x: Double) = (x * scale + offsetX)
-        fun mapY(y: Double) = (y * scale + offsetY)
+        fun mapX(x: Double) = (x * scale + offsetX).toFloat()
+        fun mapY(y: Double) = (y * scale + offsetY).toFloat()
 
         val path = Path()
         points.forEachIndexed { index, p ->
